feat(trin-execution): replace EraManager with E2HSManager
