+++ bin/e2hs-writer/Cargo.toml
@@ -27,12 +27,14 @@ futures.workspace = true
 humanize-duration.workspace = true
 reqwest.workspace = true
 rstest.workspace = true
+revm-primitives.workspace = true
 serde_yaml.workspace = true
 ssz_types.workspace = true
 tempfile.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 tree_hash.workspace = true
+trin-evm.workspace = true
 trin-execution.workspace = true
 trin-utils.workspace = true
 trin-validation.workspace = true
