+++ bin/trin-execution/src/e2hs/mod.rs
@@ -1,0 +1,4 @@
+pub mod beacon;
+pub mod manager;
+pub mod types;
+pub mod utils;
