--- bin/trin-execution/src/era/constants.rs
@@ -1,1 +1,0 @@
-pub const FIRST_ERA_EPOCH_WITH_EXECUTION_PAYLOAD: u64 = 574;
