--- bin/trin-execution/src/era/mod.rs
@@ -1,6 +1,0 @@
-pub mod beacon;
-pub mod binary_search;
-pub mod constants;
-pub mod manager;
-pub mod types;
-pub mod utils;
