+++ bin/trin-execution/src/evm/post_block_beneficiaries.rs
@@ -9,7 +9,7 @@ use revm_primitives::hardfork::SpecId;
 use trin_evm::spec_id::get_spec_block_number;
 
 use super::dao_fork::{DAO_HARDFORK_BENEFICIARY, DAO_HARDKFORK_ACCOUNTS};
-use crate::{era::types::ProcessedBlock, storage::evm_db::EvmDB};
+use crate::{e2hs::types::ProcessedBlock, storage::evm_db::EvmDB};
 
 // Calculate block reward
 // https://github.com/paradigmxyz/reth/blob/v0.2.0-beta.6/crates/consensus/common/src/calc.rs
