+++ bin/trin-execution/src/subcommands/e2ss/import.rs
@@ -10,7 +10,7 @@ use tracing::info;
 use crate::{
     cli::ImportStateConfig,
     config::StateConfig,
-    era::manager::EraManager,
+    e2hs::manager::E2HSManager,
     evm::block_executor::BLOCKHASH_SERVE_WINDOW,
     storage::{
         account_db::AccountDB, evm_db::EvmDB, execution_position::ExecutionPosition,
@@ -139,9 +139,9 @@ impl StateImporter {
     /// insert the last 256 block hashes into the database
     async fn import_last_256_block_hashes(&self, block_number: u64) -> anyhow::Result<()> {
         let first_block_hash_to_add = block_number.saturating_sub(BLOCKHASH_SERVE_WINDOW);
-        let mut era_manager = EraManager::new(first_block_hash_to_add).await?;
-        while era_manager.next_block_number() <= block_number {
-            let block = era_manager.get_next_block().await?;
+        let mut e2hs_manager = E2HSManager::new(first_block_hash_to_add).await?;
+        while e2hs_manager.next_block_number() <= block_number {
+            let block = e2hs_manager.get_next_block().await?;
             self.evm_db.db.put(
                 keccak256(B256::from(U256::from(block.header.number))),
                 block.header.hash_slow(),
